+++ winter-embed-scheduler-center/src/main/java/tech/obiteaaron/winter/embed/schedulercenter/JobContext.java
@@ -1,0 +1,4 @@
+package tech.obiteaaron.winter.embed.schedulercenter;
+
+public class JobContext {
+}