+++ winter-embed-scheduler-center/src/main/java/tech/obiteaaron/winter/embed/schedulercenter/JobProcessor.java
@@ -1,0 +1,9 @@
+package tech.obiteaaron.winter.embed.schedulercenter;
+
+/**
+ * 基础接口，用来组合使用
+ */
+public interface JobProcessor {
+
+    void process(JobContext jobContext);
+}