+++ winter-embed-scheduler-center/src/main/java/tech/obiteaaron/winter/embed/schedulercenter/LongTimeJobProcessor.java
@@ -1,0 +1,8 @@
+package tech.obiteaaron.winter.embed.schedulercenter;
+
+/**
+ * 常驻任务处理器
+ * 执行一次后不退出，直到达成退出条件
+ */
+public interface LongTimeJobProcessor extends JobProcessor {
+}