+++ winter-embed-scheduler-center/src/main/java/tech/obiteaaron/winter/embed/schedulercenter/MapReduceJobProcessor.java
@@ -1,0 +1,8 @@
+package tech.obiteaaron.winter.embed.schedulercenter;
+
+/**
+ * 多机分发任务处理器，分发后合并结果
+ * 执行一次后不退出，直到达成退出条件
+ */
+public interface MapReduceJobProcessor extends MapJobProcessor {
+}