+++ winter-embed-scheduler-center/src/main/java/tech/obiteaaron/winter/embed/schedulercenter/SimpleJobProcessor.java
@@ -1,0 +1,7 @@
+package tech.obiteaaron.winter.embed.schedulercenter;
+
+/**
+ * 简单任务处理器，执行一次后立即退出
+ */
+public interface SimpleJobProcessor extends JobProcessor {
+}