+++ winter-embed-scheduler-center/src/main/java/tech/obiteaaron/winter/embed/schedulercenter/WinterSchedulerCenterConfig.java
@@ -1,0 +1,10 @@
+package tech.obiteaaron.winter.embed.schedulercenter;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class WinterSchedulerCenterConfig {
+
+}