+++ winter-embed-scheduler-center/src/main/java/tech/obiteaaron/winter/embed/schedulercenter/spring/WinterSchedulerCenterSpringAutoConfiguration.java
@@ -1,0 +1,10 @@
+package tech.obiteaaron.winter.embed.schedulercenter.spring;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@EnableConfigurationProperties(WinterSchedulerCenterProperties.class)
+@Configuration
+public class WinterSchedulerCenterSpringAutoConfiguration {
+
+}