+++ winter-embed-scheduler-center/CHANGELOG.md
@@ -3,4 +3,4 @@
 发布日期：coding
 发布内容：
 
-- 提供嵌入式的、集群内多机MapReduce任务执行能力，依赖winter-embed-rpc
+- 提供嵌入式的、集群内多机MapReduce任务执行能力，依赖winter-embed-rpc、winter-embed-config-center